requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tencentcloud-sdk-python-common==3.0.1073
tencentcloud-sdk-python-tmt==3.0.1073
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
